Updated Factorial using BigInteger Class

+++ Factorial.java
@@ -43,8 +43,9 @@ public class Factorial{
 	 * @param n The number to factorial
 	 * @return The factorial of the number
 	 */
-	public static long factorial(int n){
+	public static BigInteger factorial(int n){
           if(n == 0 || n == 1) return 1;
           return n * factorial(n - 1);
 	}
+
 }
